+++ backend/src/routes/api/stays.js
@@ -11,44 +11,66 @@ const router = express.Router();
 // RESTful routes here
 
 router.post('/', async (req, res) => {
-  const response = await staysDao.createStay(req.body);
+  const response = await staysDao.createStay(req.body, req.user.sub);
   res.json(response);
 });
 
 router.get('/', async (req, res) => {
-  res.json(await staysDao.retrieveAllStays());
+  res.json(await staysDao.retrieveAllStays(req.user.sub));
 });
 
 router.get('/:id', async (req, res) => {
   const { id: tripId } = req.params;
-
-  try {
-    const stays = await staysDao.retrieveStays(tripId);
-
-    if (stays) {
-      res.json(stays);
-    } else {
-      res.sendStatus(HTTP_NOT_FOUND);
+  const stays = await staysDao.retrieveStays(tripId);
+  if (stays) {
+    if (stays.userSub !== req.user.sub)
+    {
+      res.json('Error Id is not valid');
     }
-  } catch (err) {
-    res.json('Error Id is not valid');
+    else
+    {
+      res.json(stays);
+    }
+  }
+  else
+  {
+    res.sendStatus(HTTP_NOT_FOUND);
   }
 });
 
 router.put('/:id', async (req, res) => {
   const { id } = req.params;
-  const response = res.json(await staysDao.updateStay(id, req.body));
-  if (response) {
-    res.sendStatus(HTTP_NO_CONTENT);
-  } else {
-    res.sendStatus(HTTP_NOT_FOUND);
+  const stayToUpdate = await staysDao.retrieveStays(id);
+  if (stayToUpdate) {
+    if (stayToUpdate.userSub !== req.user.sub)
+    {
+      res.sendStatus(400);
+    }
+    else
+    {
+      const response = res.json(await staysDao.updateStay(id, req.body));
+      res.sendStatus(response ? HTTP_NO_CONTENT : HTTP_NOT_FOUND);
+    }
+  } else
+  {
+    res.sendStatus(HTTP_NOT_FOUND)
   }
 });
 
 router.delete('/:id', async (req, res) => {
   const { id } = req.params;
-  const response = await staysDao.deleteStay(id);
-  res.json(response);
+  const stayDelete = await staysDao.retrieveStays(id);
+  if (stayDelete) {
+    if (stayDelete.userSub !== req.user.sub)
+    {
+      res.sendStatus(400);
+    } else {
+      const response = await staysDao.deleteStay(id);
+      res.json(response);
+    }
+  } else {
+    res.sendStatus(HTTP_NOT_FOUND);
+  }
 });
 
 export default router;
